+++ sdk/signalr/Microsoft.Azure.WebJobs.Extensions.SignalRService/src/Config/OptionsSetup.cs
@@ -30,6 +30,10 @@ namespace Microsoft.Azure.WebJobs.Extensions.SignalRService
                 options.ServiceTransportType = optionsFromCode.ServiceTransportType;
                 options.UseJsonObjectSerializer(optionsFromCode.JsonObjectSerializer);
                 options.RetryOptions = optionsFromCode.RetryOptions;
+                if (optionsFromCode.HttpClientTimeout.HasValue)
+                {
+                    options.HttpClientTimeout = optionsFromCode.HttpClientTimeout.Value;
+                }
 
                 // Apply options from configuration
                 if (_configuration.TryGetJsonObjectSerializer(out var serializer))
@@ -67,6 +71,12 @@ namespace Microsoft.Azure.WebJobs.Extensions.SignalRService
                 if (retryOptions != null)
                 {
                     options.RetryOptions = retryOptions;
+                }
+
+                var httpClientTimeout = _configuration.GetSection(Constants.AzureSignalRHttpClientTimeout).Get<TimeSpan>();
+                if (httpClientTimeout != default)
+                {
+                    options.HttpClientTimeout = httpClientTimeout;
                 }
             };
         }